Simplify and unit-test cmp-matrices
