+++ internal/pkg/mlrval/mlrval_constants.go
@@ -21,30 +21,6 @@ func MlrvalFromPending() Mlrval {
 // just using `const` isn't enough) but the gentelpersons' agreement is that
 // the caller should never modify these.
 
-var ERROR = &Mlrval{
-	mvtype:        MT_ERROR,
-	printrep:      ERROR_PRINTREP,
-	printrepValid: true,
-}
-
-var ABSENT = &Mlrval{
-	mvtype:        MT_ABSENT,
-	printrep:      ABSENT_PRINTREP,
-	printrepValid: true,
-}
-
-var NULL = &Mlrval{
-	mvtype:        MT_NULL,
-	printrep:      "null",
-	printrepValid: true,
-}
-
-var VOID = &Mlrval{
-	mvtype:        MT_VOID,
-	printrep:      "",
-	printrepValid: true,
-}
-
 var TRUE = &Mlrval{
 	mvtype:        MT_BOOL,
 	printrep:      "true",
@@ -56,4 +32,50 @@ var FALSE = &Mlrval{
 	mvtype:        MT_BOOL,
 	printrep:      "false",
 	printrepValid: true,
+}
+
+var VOID = &Mlrval{
+	mvtype:        MT_VOID,
+	printrep:      "",
+	printrepValid: true,
+}
+
+var ERROR = &Mlrval{
+	mvtype:        MT_ERROR,
+	printrep:      ERROR_PRINTREP,
+	printrepValid: true,
+}
+
+var NULL = &Mlrval{
+	mvtype:        MT_NULL,
+	printrep:      "null",
+	printrepValid: true,
+}
+
+var ABSENT = &Mlrval{
+	mvtype:        MT_ABSENT,
+	printrep:      ABSENT_PRINTREP,
+	printrepValid: true,
+}
+
+// For malloc-avoidance in the spaceship operator
+var MINUS_ONE = &Mlrval{
+	mvtype:        MT_INT,
+	printrep:      "-1",
+	printrepValid: true,
+	intval:        -1,
+}
+
+var ZERO = &Mlrval{
+	mvtype:        MT_INT,
+	printrep:      "0",
+	printrepValid: true,
+	intval:        0,
+}
+
+var ONE = &Mlrval{
+	mvtype:        MT_INT,
+	printrep:      "1",
+	printrepValid: true,
+	intval:        1,
 }
