+++ test/cases/mix-null-field-input/0021/expout
@@ -1,6 +1,6 @@
 top_idx x_top
-1       5
-2       3
-3       1
-4       -
-5       -
+1       -
+2       -
+3       5
+4       3
+5       1
