+++ test/cases/mix-null-field-input/0022/expout
@@ -1,6 +1,6 @@
 top_idx y_top
-1       6
-2       4
-3       2
-4       -
-5       -
+1       -
+2       -
+3       6
+4       4
+5       2
