+++ test/cases/mix-null-field-input/0024/expout
@@ -1,6 +1,6 @@
 top_idx x_top y_top z_top
-1       5     6     -
-2       3     4     -
-3       1     2     -
-4       -     -     -
-5       -     -     -
+1       -     -     -
+2       -     -     -
+3       5     6     -
+4       3     4     -
+5       1     2     -
