+++ test/cases/mix-null-field-input/0030/expout
@@ -1,6 +1,6 @@
 a x y z
+v - - -
+t 5 - -
 u - 6 -
 s 3 4 -
 r 1 2 -
-v - - -
-t 5 - -
