+++ test/cases/mix-null-field-input/0037/expout
@@ -1,6 +1,6 @@
 a x y z
+v - - -
+u - 6 -
 t 5 - -
 s 3 4 -
 r 1 2 -
-v - - -
-u - 6 -
